requirements.txt
beautifulsoup4==4.9.0
Flask==1.1.2
Flask-Compress==1.4.0
Flask-WTF==0.14.3
IMDbPY==6.8
more-itertools==8.2.0
numpy==1.18.4
numpydoc==0.9.2
pandas==1.0.3
requests==2.23.0
WTForms==2.3.1